Prevent deleting records that are still in use

+++ modules/employees/delete.php
@@ -1,6 +1,7 @@
 <?php
 require '../../config/config.php';
 require '../../includes/employee_system_access.php';
+
 esa_ensure_table($conn);
 
 $id = 0;
@@ -11,10 +12,15 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 }
 
 if ($id > 0) {
-    mysqli_query($conn, 'DELETE FROM employee_system_access_relations WHERE employee_id=' . $id . ' AND company_id=' . (int)$company_id);
-    mysqli_query($conn, 'DELETE FROM employee_system_access WHERE employee_id=' . $id . ' AND company_id=' . (int)$company_id);
-    $sql = 'DELETE FROM employees WHERE id=' . $id . ' AND company_id=' . (int)$company_id . ' LIMIT 1';
-    mysqli_query($conn, $sql);
+    $usageError = '';
+    if (!itm_can_delete_record($conn, 'employees', 'id', $id, $company_id, $usageError)) {
+        $_SESSION['crud_error'] = $usageError;
+    } else {
+        mysqli_query($conn, 'DELETE FROM employee_system_access_relations WHERE employee_id=' . $id . ' AND company_id=' . (int)$company_id);
+        mysqli_query($conn, 'DELETE FROM employee_system_access WHERE employee_id=' . $id . ' AND company_id=' . (int)$company_id);
+        $sql = 'DELETE FROM employees WHERE id=' . $id . ' AND company_id=' . (int)$company_id . ' LIMIT 1';
+        mysqli_query($conn, $sql);
+    }
 }
 
 header('Location: index.php');
