+++ modules/system_access/delete.php
@@ -6,7 +6,13 @@ esa_ensure_table($conn);
 
 $id = (int)($_GET['id'] ?? 0);
 if ($id > 0) {
-    mysqli_query($conn, "DELETE FROM system_access WHERE id=$id AND company_id=$company_id");
+    $usageError = '';
+    if (!itm_can_delete_record($conn, 'system_access', 'id', $id, $company_id, $usageError)) {
+        $_SESSION['crud_error'] = $usageError;
+    } else {
+        mysqli_query($conn, "DELETE FROM system_access WHERE id=$id AND company_id=$company_id");
+    }
 }
+
 header('Location: index.php');
 exit;
